# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for MNI Higher Secondary School
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  isAdmin: boolean;
  username?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  titleHindi: string;
  content: string;
  contentHindi: string;
  author: string;
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBlogPostBody {
  title: string;
  titleHindi: string;
  content: string;
  contentHindi: string;
  author: string;
  imageUrl?: string | null;
}

export interface GalleryPhoto {
  id: number;
  title: string;
  titleHindi: string;
  imageUrl: string;
  category: string;
  createdAt: string;
}

export interface CreateGalleryPhotoBody {
  title: string;
  titleHindi: string;
  imageUrl: string;
  category: string;
}

export interface StaffMember {
  id: number;
  name: string;
  nameHindi: string;
  role: string;
  roleHindi: string;
  photoUrl?: string | null;
  order: number;
  createdAt: string;
}

export interface CreateStaffBody {
  name: string;
  nameHindi: string;
  role: string;
  roleHindi: string;
  photoUrl?: string | null;
  order: number;
}