# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for MNI Higher Secondary School
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get current admin session
 */
export const GetAdminMeResponse = zod.object({
  isAdmin: zod.boolean(),
  username: zod.string().optional(),
});

/**
 * @summary List all blog posts
 */
export const ListBlogPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  titleHindi: zod.string(),
  content: zod.string(),
  contentHindi: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBlogPostsResponse = zod.array(ListBlogPostsResponseItem);

/**
 * @summary Create a blog post
 */
export const CreateBlogPostBody = zod.object({
  title: zod.string(),
  titleHindi: zod.string(),
  content: zod.string(),
  contentHindi: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a blog post by ID
 */
export const GetBlogPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBlogPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  titleHindi: zod.string(),
  content: zod.string(),
  contentHindi: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a blog post
 */
export const UpdateBlogPostParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBlogPostBody = zod.object({
  title: zod.string(),
  titleHindi: zod.string(),
  content: zod.string(),
  contentHindi: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
});

export const UpdateBlogPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  titleHindi: zod.string(),
  content: zod.string(),
  contentHindi: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a blog post
 */
export const DeleteBlogPostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all gallery photos
 */
export const ListGalleryPhotosResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  titleHindi: zod.string(),
  imageUrl: zod.string(),
  category: zod.string(),
  createdAt: zod.string(),
});
export const ListGalleryPhotosResponse = zod.array(
  ListGalleryPhotosResponseItem,
);

/**
 * @summary Add a photo to gallery
 */
export const CreateGalleryPhotoBody = zod.object({
  title: zod.string(),
  titleHindi: zod.string(),
  imageUrl: zod.string(),
  category: zod.string(),
});

/**
 * @summary Update gallery photo
 */
export const UpdateGalleryPhotoParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateGalleryPhotoBody = zod.object({
  title: zod.string(),
  titleHindi: zod.string(),
  imageUrl: zod.string(),
  category: zod.string(),
});

export const UpdateGalleryPhotoResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  titleHindi: zod.string(),
  imageUrl: zod.string(),
  category: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a gallery photo
 */
export const DeleteGalleryPhotoParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all administrative staff
 */
export const ListStaffResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameHindi: zod.string(),
  role: zod.string(),
  roleHindi: zod.string(),
  photoUrl: zod.string().nullish(),
  order: zod.number(),
  createdAt: zod.string(),
});
export const ListStaffResponse = zod.array(ListStaffResponseItem);

/**
 * @summary Add a staff member
 */
export const CreateStaffBody = zod.object({
  name: zod.string(),
  nameHindi: zod.string(),
  role: zod.string(),
  roleHindi: zod.string(),
  photoUrl: zod.string().nullish(),
  order: zod.number(),
});

/**
 * @summary Update a staff member
 */
export const UpdateStaffParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateStaffBody = zod.object({
  name: zod.string(),
  nameHindi: zod.string(),
  role: zod.string(),
  roleHindi: zod.string(),
  photoUrl: zod.string().nullish(),
  order: zod.number(),
});

export const UpdateStaffResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameHindi: zod.string(),
  role: zod.string(),
  roleHindi: zod.string(),
  photoUrl: zod.string().nullish(),
  order: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a staff member
 */
export const DeleteStaffParams = zod.object({
  id: zod.coerce.number(),
});