# lib/api-zod/src/generated/types/createBlogPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for MNI Higher Secondary School
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBlogPostBody {
  title: string;
  titleHindi: string;
  content: string;
  contentHindi: string;
  author: string;
  imageUrl?: string | null;
}