# lib/api-zod/src/generated/types/galleryPhoto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for MNI Higher Secondary School
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryPhoto {
  id: number;
  title: string;
  titleHindi: string;
  imageUrl: string;
  category: string;
  createdAt: string;
}